--- docs/examples/chess/.dlt/config.toml
@@ -1,9 +1,0 @@
-chess_url="https://api.chess.com/pub/"
-pipeline_name="chess"
-
-[runtime]
-log_level="WARNING"
-
-[sources.chess.extract]
-max_parallel_items = 5
-workers = 2
