+++ docs/examples/pdf_to_weaviate/assets/invoices/invoice_2.txt
