+++ docs/examples/pdf_to_weaviate/assets/invoices/invoice_20230831-p1.pdf
