--- docs/website/docs/examples/custom_destination_bigquery/code/.dlt/example.secrets.toml
@@ -1,10 +1,0 @@
-# @@@DLT_SNIPPET_START example
-[destination.bigquery.credentials]
-client_email = ""
-private_key = ""
-project_id = ""
-token_uri = ""
-refresh_token = ""
-client_id = ""
-client_secret = ""
-# @@@DLT_SNIPPET_END example
