--- docs/website/docs/examples/google_sheets/code/.dlt/config.toml
@@ -1,2 +1,0 @@
-# @@@DLT_SNIPPET_START example
-# @@@DLT_SNIPPET_END example
