--- docs/website/docs/examples/google_sheets/code/.dlt/example.secrets.toml
@@ -1,18 +1,0 @@
-# @@@DLT_SNIPPET_START example
-# you can just paste services.json as credentials
-[sources.google_sheets]
-credentials='''
-{
-  "type": "set me up!",
-  "project_id": "set me up!",
-  "private_key_id": "set me up!",
-  "private_key": "set me up!",
-  "client_email": "set me up!",
-  "client_id": "set me up!",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "set me up!"
-  }
-'''
-# @@@DLT_SNIPPET_END example
