--- docs/website/docs/examples/incremental_loading/code/.dlt/example.secrets.toml
@@ -1,6 +1,0 @@
-# @@@DLT_SNIPPET_START example
-[sources.zendesk.credentials]
-password = ""
-subdomain = ""
-email = ""
-# @@@DLT_SNIPPET_END example
