--- docs/website/docs/examples/nested_data/code/.dlt/config.toml
@@ -1,2 +1,0 @@
-# @@@DLT_SNIPPET_START example
-# @@@DLT_SNIPPET_END example
