--- docs/website/docs/examples/nested_data/code/.dlt/example.secrets.toml
@@ -1,4 +1,0 @@
-# @@@DLT_SNIPPET_START example
-[sources.mongodb]
-connection_url=""
-# @@@DLT_SNIPPET_END example
