--- docs/website/docs/examples/qdrant_zendesk/code/.dlt/example.secrets.toml
@@ -1,10 +1,0 @@
-# @@@DLT_SNIPPET_START example
-[destination.qdrant.credentials]
-location = ""
-api_key = ""
-
-[sources.zendesk.credentials]
-password = ""
-subdomain = ""
-email = ""
-# @@@DLT_SNIPPET_END example
