--- docs/website/docs/examples/transformers/code/.dlt/config.toml
@@ -1,18 +1,0 @@
-# @@@DLT_SNIPPET_START example
-[runtime]
-log_level="WARNING"
-
-[extract]
-# use 2 workers to extract sources in parallel
-worker=2
-# allow 10 async items to be processed in parallel
-max_parallel_items=10
-
-[normalize]
-# use 3 worker processes to process 3 files in parallel
-workers=3
-
-[load]
-# have 50 concurrent load jobs
-workers=50
-# @@@DLT_SNIPPET_END example
